Changes: Structured pages and a refined presentation layer.
