+++ lib/presentation/contact/contact.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:portfolio/core/extensions/widget_extensions.dart';
+import 'package:portfolio/presentation/contact/widgets/media_links.dart';
+
+import '../../../core/utils/device/device_config.dart';
+import '../../core/constants/string_constants.dart';
+import '../home/widgets/footer.dart';
+
+enum MediaLinkE {
+  medium(title: StringC.medium, url: StringC.mediumUrl),
+  email(title: StringC.email, url: StringC.emailUrl),
+  linkedIn(title: StringC.linkedIn, url: StringC.linkedInUrl),
+  stackOverflow(title: StringC.stackOverflow, url: StringC.stackOverflowUrl);
+  // phone(title: 'Phone', url: '+91 9924190188');
+
+  final String title;
+  final String url;
+
+  const MediaLinkE({required this.title, required this.url});
+}
+
+class Contact extends StatelessWidget {
+  const Contact({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: Device().width,
+      height: Device().height,
+      child: Column(
+        children: [
+          const MediaLinks().center.expand(),
+          const Footer(),
+        ],
+      ),
+    );
+  }
+}
